Déplacement de la fonction du champ subMenu dans les helpers de la classe User.
Passe l'utilisateur actuel dans une variable dans arborescenceStrucure.js
